Add ValueWithMod and BaseModifier classes to ModifiableValue namespace
The new ValueWithMod class in the ModifiableValue namespace adds functionality for creating values that can be altered by Modifiers. Alongside this, the BaseModifier class has been implemented to provide the Modifier objects used in the ValueWithMod class. This update is versioned as 1.0.5 in the package.json and CHANGELOG.md files.