Adjust guid length condition in ScriptableObjectExtensions
The condition to check if more than one guid is present in the ScriptableObjectExtensions file has been updated. The previous condition was checking if there was at least one guid, but it has been modified to check for more than one guid. This change optimizes the search process when multiple guids exist.